implement fetch_map
