+++ spec/unit/select_where_spec.rb
@@ -1,53 +1,24 @@
 require 'spec_helper'
 require 'select_where'
 
-RSpec.describe SelectWhere::Helpers do
+RSpec.describe SelectWhere do
 
-  describe '::item_matches_target?' do
+  it 'should implement fetch_map (key)' do
+    arr = [{ 'a' => 1 }, { 'a' => 2 }, { 'b' => 3}]
 
-    it 'should work with a symbol keyed hash' do
-      expect(SelectWhere::Helpers.item_matches_target?({ a: 10 }, 'a', 10)).to eq(true)
-      expect(SelectWhere::Helpers.item_matches_target?({ a: 10 }, :a, 10)).to eq(true)
-    end
-
-    it 'should work with a string keyed hash' do
-      expect(SelectWhere::Helpers.item_matches_target?({ 'a' => 10 }, 'a', 10)).to eq(true)
-      expect(SelectWhere::Helpers.item_matches_target?({ 'a' => 10 }, :a, 10)).to eq(true)
-    end
-
-    it 'should return false for a non-object or proc' do
-      expect(SelectWhere::Helpers.item_matches_target?(->() { 10 }, 'a', 10)).to eq(false)
-      expect(SelectWhere::Helpers.item_matches_target?(Proc.new { 10 }, 'a', 10)).to eq(false)
-    end
-
-    context 'value is a Proc' do
-
-      it 'call the proc with the found value' do
-        test_proc = ->(a) { a == 10 }
-        expect(SelectWhere::Helpers.item_matches_target?({ a: 10 }, 'a', test_proc)).to eq(true)
-        expect(SelectWhere::Helpers.item_matches_target?({ 'a' => 10 }, 'a', test_proc)).to eq(true)
-      end
-
-    end
-
+    expect {
+      SelectWhere.fetch_map(arr, 'a')
+    }.to raise_error(KeyError)
   end
 
-  describe '::value_matches_target?' do
+  it 'should implement fetch_map (key, value)' do
+    arr = [{ 'a' => 1 }, { 'a' => 2 }, { 'b' => 3}]
+    expect(SelectWhere.fetch_map(arr, 'a', 'default')).to eq([1, 2, 'default'])
+  end
 
-    it 'should work with simple equality' do
-      expect(SelectWhere::Helpers.value_matches_target?('a', 'a')).to eq(true)
-    end
-
-    it 'should work with proc values' do
-      value = -> (v) { v == 'a' }
-      expect(SelectWhere::Helpers.value_matches_target?('a', value)).to eq(true)
-    end
-
-    it 'should raise an exception if ' do
-      value = -> (v) { v == 'a' }
-      expect(SelectWhere::Helpers.value_matches_target?('a', value)).to eq(true)
-    end
-
+  it 'should implement fetch_map (key, block)' do
+    arr = [{ 'a' => 1 }, { 'a' => 2 }, { 'b' => 3}]
+    expect(SelectWhere.fetch_map(arr, 'a') { 'default' }).to eq([1, 2, 'default'])
   end
 
 end
